deadlock 탐지 관련 코드 LockManager에 추가

+++ Common/Buffer/BufReader.h
@@ -17,7 +17,7 @@ public:
 	}
 
 	BufReader& operator>>(OUT string& _val) {
-		UInt16 strLen = 0;
+		size_t strLen = 0;
 		buf.ReadBytes(reinterpret_cast<BYTE*>(&strLen), sizeof(strLen));
 		for (int idx = 0; idx < strLen; idx++) {
 			char ch = '\0';
