+++ Common/Buffer/BufWriter.h
@@ -20,7 +20,7 @@ public:
 	}
 
 	BufWriter& operator<< (string _str) {
-		UInt16 len = _str.size();
+		size_t len = _str.size();
 		buf.WriteBytes(reinterpret_cast<BYTE*>(&len), sizeof(len));
 		for (char _ch : _str) {
 			buf.WriteBytes(reinterpret_cast<BYTE*>(&_ch), sizeof(_ch));
